+++ pkg/envoy/http_connection_manager.go
@@ -26,6 +26,7 @@ import (
 	"github.com/envoyproxy/go-control-plane/pkg/wellknown"
 	"github.com/golang/protobuf/ptypes"
 	"github.com/golang/protobuf/ptypes/duration"
+	"knative.dev/net-kourier/pkg/config"
 )
 
 func NewHTTPConnectionManager(virtualHosts []*route.VirtualHost) httpconnectionmanagerv2.HttpConnectionManager {
@@ -33,7 +34,7 @@ func NewHTTPConnectionManager(virtualHosts []*route.VirtualHost) httpconnectionm
 	var filters []*httpconnectionmanagerv2.HttpFilter
 
 	// Get the extAuthzConf from envs vars.
-	extAuthzConf := GetExternalAuthzConfig()
+	extAuthzConf := config.GetExternalAuthzConfig()
 
 	if extAuthzConf.Enabled {
 		filters = append(filters, extAuthzConf.HTTPFilter)
