+++ pkg/generator/caches.go
@@ -28,6 +28,7 @@ import (
 	"github.com/google/uuid"
 	"go.uber.org/zap"
 	kubeclient "k8s.io/client-go/kubernetes"
+	"knative.dev/net-kourier/pkg/config"
 	"knative.dev/net-kourier/pkg/envoy"
 	"knative.dev/networking/pkg/apis/networking/v1alpha1"
 )
@@ -100,7 +101,7 @@ func (caches *Caches) UpdateIngress(ctx context.Context, ingress *v1alpha1.Ingre
 
 func (caches *Caches) initConfig(ctx context.Context, kubernetesClient kubeclient.Interface, extAuthz bool) error {
 	if extAuthz {
-		extAuthZConfig := envoy.GetExternalAuthzConfig()
+		extAuthZConfig := config.GetExternalAuthzConfig()
 		caches.addClusterForIngress(extAuthZConfig.Cluster, "__extAuthZCluster", "_internal")
 	}
 	caches.addStatusVirtualHost()
